# Fix memory consumption of into sqlite