--- a/crates/nu-command/tests/commands/database/mod.rs
+++ b/crates/nu-command/tests/commands/database/mod.rs
@@ -1,0 +1,1 @@
+mod into_sqlite;