--- a/crates/nu-command/tests/commands/mod.rs
+++ b/crates/nu-command/tests/commands/mod.rs
@@ -13,6 +13,7 @@ mod config_nu_default;
 mod continue_;
 mod conversions;
 mod cp;
+mod database;
 mod date;
 mod debug_info;
 mod def;