@amtoine 2023-09-10
i wanted to run the script and see the time improvement, but it looks like
- `update cells` is a command in *extra*
- `time` is not a command on my system

:confused: