@dead10ck 2023-09-10
>     * `update cells` is a command in _extra_

Yup.

>     * `time` is not a command on my system

This is the GNU time program, it's available on most package managers.