@rgwood 2023-09-10
This looks really great, thank you. We've got a lot of places where we collect when we could stream; love seeing people fix those up.