@dead10ck 2023-10-06
Ok I've added some tests. I found a couple of preexisting bugs along the way that I've worked around for now, and added explanatory comments.