@dead10ck 2023-10-12
> What kind of bugs did you come across with your testing?

I put them both in comments:

* https://github.com/nushell/nushell/issues/9186 - in one test, I generate random rows and write it out to a nuon file first so that we actually hit the streaming code path. But it is currently impossible to deserialize a nuon record with a date object in it.
* https://github.com/nushell/nushell/issues/10612 - using floats to parse file sizes leads to incorrectly parsed literal integral values when the value is large enough to lose precision. I'm working around this right now by just limiting the range of the randomly generated file size