@dead10ck 2023-11-15
Ok I've updated #10692 , which this is based on, to remove the changes to the `nu!` macro